move the component into a better location change the name of the searchPage property
